Add authentication with Google and GitHub
This commit implements user authentication via Google and GitHub. The authentication flow is defined in the 'src/hooks.server.ts' file and sign-in and sign-out functions are added to the login page. Login screen layout and assets are also established. Additionally, Redirection to the "/login" route in the admin layout is modified. Finally, Github and Google SVG icons are introduced to the lib components.