Refactor readme, update server code, and enhance auth configuration
This commit includes instructions simplification in the README file. In
the server code, in src/hooks.server.ts, the auth configuration is
updated where multiple environment variables are merged into a single
command. Google and GitHub authentication is enhanced by updating client
secret values and adding auth secret. Additionally, Sveltekit's
adapter-node is added to the package.json dependencies list to inform
the target environment. The server.js is a new file for handling the
server-side related activities, and it uses dotenv for managing
environment variables.